chore: move config files
