--- linaria.config.js
@@ -1,5 +1,0 @@
-// @FIXME should be in the same dir as webpackconfig
-module.exports = {
-  evaluate: true,
-  displayName: process.env.NODE_ENV !== 'production',
-};
