+++ packages/configs/babel.config.js
@@ -1,0 +1,20 @@
+module.exports = {
+  presets: [
+    '@babel/preset-env',
+    '@babel/preset-typescript',
+    [
+      '@babel/preset-react',
+      {
+        runtime: 'automatic',
+        // @TODO check out what this does
+       // development: process.env.BABEL_ENV === 'development',
+      },
+    ],
+    '@linaria',
+  ],
+  plugins: [
+    ['@babel/plugin-proposal-private-methods', { loose: true }],
+    ['@babel/plugin-proposal-class-properties', { loose: true }],
+    ['@babel/plugin-proposal-private-property-in-object', { loose: true }],
+  ],
+};
