+++ packages/configs/webpack.config.ts
@@ -1,0 +1,298 @@
+import ReactRefreshWebpackPlugin from '@pmmmwh/react-refresh-webpack-plugin';
+import CompressionPlugin from 'compression-webpack-plugin';
+import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
+import DotEnv from 'dotenv-webpack';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+import ImageMinimizerPlugin from 'image-minimizer-webpack-plugin';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+import path from 'path';
+import TerserPlugin from 'terser-webpack-plugin';
+import type { Configuration as WebpackConfiguration, WebpackPluginInstance } from 'webpack';
+import webpack from 'webpack';
+import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
+import type { Configuration as WebpackDevServerConfiguration } from 'webpack-dev-server';
+import zlib from 'zlib';
+
+interface Configuration extends WebpackConfiguration, WebpackDevServerConfiguration {
+  devServer?: WebpackDevServerConfiguration;
+}
+
+type CustomEnv = {
+  analyze: boolean;
+  name: string;
+};
+
+type ArgV = {
+  mode: Configuration['mode'];
+};
+
+type ConfigFn = (env: CustomEnv, argv: ArgV) => Configuration;
+
+const MAX_CHUNK_SIZE = 300000;
+const DEV_PORT = 9000;
+const APP_TITLE = 'Solana Wallet';
+const WEB_PATH = path.resolve(__dirname, '../web/');
+const PACKAGES_PATH = path.resolve(__dirname, '../');
+
+const config: ConfigFn = (env, argv) => {
+  const __DEVELOPMENT__ = argv.mode === 'development';
+  const __PRODUCTION__ = argv.mode === 'production';
+  const __ANALYSE__ = env.analyze;
+  const __PACKAGE_NAME__ = env.name;
+
+  // @ts-ignore
+  process.env.NODE_ENV = argv.mode;
+  process.env.BABEL_ENV = argv.mode;
+
+  const devPlugins: Array<WebpackPluginInstance> = [];
+  const utilityPlugins: Array<WebpackPluginInstance> = [];
+  const prodPlugins: Array<WebpackPluginInstance> = [];
+
+  const webAliases = {
+    constants: path.resolve(PACKAGES_PATH, '/constants'),
+    config: path.resolve(PACKAGES_PATH, '/config'),
+    utils: path.resolve(PACKAGES_PATH, '/utils'),
+    lib: path.resolve(PACKAGES_PATH, '/lib'),
+    store: path.resolve(PACKAGES_PATH, '/store'),
+    api: path.resolve(PACKAGES_PATH, '/api'),
+    app: path.resolve(PACKAGES_PATH, '/app'),
+    pages: path.resolve(PACKAGES_PATH, '/pages'),
+    components: path.resolve(PACKAGES_PATH, '/components'),
+    assets: path.resolve(PACKAGES_PATH, '/assets'),
+    styles: path.resolve(PACKAGES_PATH, '/styles'),
+  };
+
+  const packageAliases: object = {
+    '@p2p-wallet-web/core': path.resolve(PACKAGES_PATH, '/core/src'),
+    '@p2p-wallet-web/sail': path.resolve(PACKAGES_PATH, '/sail/src'),
+    '@p2p-wallet-web/token-utils': path.resolve(PACKAGES_PATH, '/token-utils/src'),
+    '@p2p-wallet-web/ui': path.resolve(PACKAGES_PATH, '/ui/src'),
+  };
+
+  if (__PRODUCTION__) {
+    prodPlugins.push(
+      new CompressionPlugin({
+        filename: '[path][base].br',
+        algorithm: 'brotliCompress',
+        test: /\.(js|css|svg)$/,
+        compressionOptions: {
+          // @ts-ignore
+          params: {
+            [zlib.constants.BROTLI_PARAM_QUALITY]: 11,
+          },
+        },
+        threshold: 10240,
+        minRatio: 0.8,
+      }),
+      new webpack.optimize.LimitChunkCountPlugin({
+        maxChunks: 10,
+      }),
+      new MiniCssExtractPlugin({
+        filename: 'styles-[contenthash].css',
+        ignoreOrder: true,
+      }),
+    );
+  }
+
+  if (__DEVELOPMENT__) {
+    devPlugins.push(new ReactRefreshWebpackPlugin());
+  }
+
+  // @TODO try out the --analize flag
+  if (__ANALYSE__) {
+    utilityPlugins.push(new BundleAnalyzerPlugin());
+  }
+
+  return {
+    bail: __PRODUCTION__,
+
+    name: __PACKAGE_NAME__,
+
+    mode: argv.mode,
+
+    entry: path.resolve(__dirname, WEB_PATH, 'src/index.tsx'),
+
+    // @TODO output for library https://webpack.js.org/configuration/output/#outputlibrary
+    // @TODO https://webpack.js.org/guides/author-libraries/
+    // set up sideeffect as well for the libs
+    output: {
+      path: path.resolve(__dirname, WEB_PATH, 'public'),
+      filename: __DEVELOPMENT__ ? '[name].[contenthash].js' : '[contenthash].js',
+      chunkFilename: __DEVELOPMENT__ ? '[id]-[contenthash].chunk.js' : '[contenthash].chunk.js',
+      assetModuleFilename: '[name]-[contenthash][ext]',
+    },
+
+    optimization: {
+      // nodeEnv: argv.mode,
+      runtimeChunk: 'single',
+      minimize: __PRODUCTION__,
+      minimizer: __PRODUCTION__
+        ? [
+            new TerserPlugin(),
+            new CssMinimizerPlugin(),
+            new ImageMinimizerPlugin({
+              loader: false,
+              minimizer: {
+                implementation: ImageMinimizerPlugin.imageminMinify,
+                options: {
+                  plugins: [
+                    'imagemin-gifsicle',
+                    'imagemin-mozjpeg',
+                    'imagemin-pngquant',
+                    'imagemin-svgo',
+                  ],
+                },
+              },
+            }),
+          ]
+        : [],
+      splitChunks: {
+        maxSize: __PRODUCTION__ ? MAX_CHUNK_SIZE : undefined,
+        cacheGroups: {
+          vendor: {
+            test: /[\\/]node_modules[\\/]/,
+            name: 'vendor',
+            chunks: 'all',
+            reuseExistingChunk: true,
+          },
+        },
+      },
+    },
+
+    performance: {
+      hints: 'warning',
+    },
+
+    module: {
+      rules: [
+        {
+          test: /.(png)$/,
+          type: 'asset/resource',
+        },
+        {
+          test: /-icon\.svg$/,
+          use: [
+            {
+              loader: 'svg-sprite-loader',
+            },
+            {
+              loader: 'svgo-loader',
+            },
+          ],
+        },
+        {
+          test: /\.css$/,
+          use: [
+            {
+              loader: __DEVELOPMENT__ ? 'style-loader' : MiniCssExtractPlugin.loader,
+            },
+            {
+              loader: 'css-loader',
+              options: {
+                sourceMap: __DEVELOPMENT__,
+                // @TODO in the case of css-modules use getLocalIndent
+              },
+            },
+          ],
+        },
+        {
+          test: /\.(ts|js)x?$/i,
+          exclude: /node_modules/,
+          use: [
+            {
+              loader: 'babel-loader',
+            },
+            {
+              loader: '@linaria/webpack-loader',
+              options: {
+                sourceMap: __DEVELOPMENT__,
+              },
+            },
+          ],
+        },
+      ],
+    },
+
+    resolve: {
+      extensions: ['.ts', '.tsx', '.js', '.jsx', '.json', '.css'],
+      fallback: {
+        http: require.resolve('stream-http'),
+        https: require.resolve('https-browserify'),
+        crypto: require.resolve('crypto-browserify'),
+        stream: require.resolve('stream-browserify'),
+        Buffer: require.resolve('buffer'),
+      },
+      alias: {
+        ...webAliases,
+        ...packageAliases,
+      },
+    },
+
+    devtool: __DEVELOPMENT__ ? 'eval-cheap-module-source-map' : 'source-map',
+
+    devServer: {
+      client: {
+        overlay: {
+          errors: true,
+          warnings: false,
+        },
+        progress: true,
+      },
+      historyApiFallback: true,
+      compress: true,
+      port: DEV_PORT,
+      hot: true,
+      open: false,
+    },
+
+    stats: __PRODUCTION__ ? 'errors-only' : 'normal',
+
+    target: 'browserslist',
+
+    plugins: [
+      new HtmlWebpackPlugin(
+        Object.assign(
+          {},
+
+          {
+            title: APP_TITLE,
+            template: path.join(__dirname + '/index.html'),
+          },
+          __PRODUCTION__ && {
+            minify: {
+              removeComments: true,
+              collapseWhitespace: true,
+              removeRedundantAttributes: true,
+              useShortDoctype: true,
+              removeEmptyAttributes: true,
+              removeStyleLinkTypeAttributes: true,
+              keepClosingSlash: true,
+              minifyJS: true,
+              minifyCSS: true,
+              minifyURLs: true,
+            },
+          },
+        ),
+      ),
+      new DotEnv({
+        // path: './packages/web/.env.development',
+        path: path.resolve(WEB_PATH, '.env.development'),
+        ignoreStub: true,
+      }),
+      new webpack.DefinePlugin({
+        __DEVELOPMENT__,
+        __PRODUCTION__,
+      }),
+      new webpack.ProvidePlugin({
+        process: 'process/browser.js',
+        Buffer: ['buffer', 'Buffer'],
+      }),
+      new webpack.ProgressPlugin(),
+      ...utilityPlugins,
+      ...devPlugins,
+      ...prodPlugins,
+    ],
+  };
+};
+
+module.exports = config;
